vscode snippets (zettelkasten), @@ctrl+space
Wow, that was a quick hack to get up-to-date autocompletion for all
internal links. By using the permalinks, we really should not have to
worry about links changing (which is the anti-pattern horror of markdown
links in blogs - we link to an ID of some kind, not a foreign
key/embedded psot).